+++ soxs/tests/test_thermal.py
@@ -2,15 +2,15 @@ import numpy as np
 import os
 import shutil
 import tempfile
-from soxs.spectra import ApecGenerator
+from soxs.spectra import ApecGenerator, get_wabs_absorb
 from soxs.spatial import PointSourceModel
 from soxs.simput import write_photon_list
 from soxs.instrument_registry import \
     get_instrument_from_registry
 from soxs.instrument import instrument_simulator, \
     RedistributionMatrixFile, AuxiliaryResponseFile
-from soxs.tests.utils import write_spectrum, get_wabs_absorb, \
-    convert_rmf
+from soxs.utils import convert_rmf
+from soxs.events import write_spectrum
 from sherpa.astro.ui import load_user_model, add_user_pars, \
     load_pha, ignore, fit, set_model, set_stat, set_method, \
     covar, get_covar_results, set_covar_opt
@@ -27,7 +27,7 @@ def mymodel(pars, x, xhi=None):
     dx = x[1]-x[0]
     wabs = get_wabs_absorb(x+0.5*dx, pars[0])
     apec = agen.get_spectrum(pars[1], pars[2], pars[3], pars[4])
-    eidxs = np.logical_and(rmf.elo >= x[0]-0.5*dx, rmf.elo-0.5*dx <= x[-1])
+    eidxs = np.logical_and(rmf.elo >= x[0]-0.5*dx, rmf.elo <= x[-1]+0.5*dx)
     return dx*wabs*apec.flux.value[eidxs]
 
 def test_thermal():